+++ api/src/main/java/com/example/muse/domain/auth/userInfo/GoogleUserInfo.java
@@ -1,16 +1,14 @@
-package com.example.muse.domain.auth;
+package com.example.muse.domain.auth.userInfo;
 
 import com.example.muse.domain.member.Provider;
-import org.springframework.security.oauth2.core.oidc.user.OidcUser;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Component;
 
-@Component("kakao")
-public class KaKaoUserInfo implements OAuth2UserInfo {
-
+@Component("google")
+public class GoogleUserInfo implements OAuth2UserInfo {
     @Override
     public Provider getProvider() {
-        return Provider.KAKAO;
+        return Provider.GOOGLE;
     }
 
     @Override
@@ -20,7 +18,6 @@ public class KaKaoUserInfo implements OAuth2UserInfo {
 
     @Override
     public String getNickname(OAuth2User user) {
-
-        return user.getAttribute("nickname");
+        return user.getAttribute("name");
     }
 }
