+++ api/src/main/java/com/example/muse/domain/auth/userInfo/KaKaoUserInfo.java
@@ -1,15 +1,15 @@
-package com.example.muse.domain.auth;
+package com.example.muse.domain.auth.userInfo;
 
 import com.example.muse.domain.member.Provider;
-import org.springframework.security.oauth2.core.oidc.user.OidcUser;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Component;
 
-@Component("google")
-public class GoogleUserInfo implements OAuth2UserInfo {
+@Component("kakao")
+public class KaKaoUserInfo implements OAuth2UserInfo {
+
     @Override
     public Provider getProvider() {
-        return Provider.GOOGLE;
+        return Provider.KAKAO;
     }
 
     @Override
@@ -19,6 +19,7 @@ public class GoogleUserInfo implements OAuth2UserInfo {
 
     @Override
     public String getNickname(OAuth2User user) {
-        return user.getAttribute("name");
+
+        return user.getAttribute("nickname");
     }
 }
